requirements.txt
requests==2.32.3
pyyaml==6.0.2
numpy==2.1.3
prometheus-api-client==0.5.5
